refactor variable_sources/utils to util for parity with top level util pkg
