+++ internal/resolution/variable_sources/util/predicates/predicates.go
