+++ internal/resolution/variable_sources/util/sort/sort.go
